Refactor NetworkTest
